Fix storage types and update tsconfig
