+++ tsconfig.json
@@ -7,6 +7,8 @@
     "noEmit": true,
     "module": "ESNext",
     "strict": true,
+    "target": "es2015",
+    "downlevelIteration": true,
     "lib": ["esnext", "dom", "dom.iterable"],
     "jsx": "preserve",
     "esModuleInterop": true,
